some improvment

+++ contracts/genesisTree/IGenesisTree.sol
@@ -31,9 +31,9 @@ interface IGenesisTree {
 
     function verifyUpdate(uint256 treeId, bool isVerified) external;
 
-    function checkAndSetProvideStatus(uint256 treeId, uint8 provideType)
+    function checkAndSetProvideStatus(uint256 treeId, uint16 provideType)
         external
-        returns (uint8);
+        returns (uint16);
 
     function updateOwner(uint256 treeId, address ownerId) external;
